add lat/lon topLeft and bottomRight to Cell as double[]

+++ src/main/java/at/ac/ait/ubicity/truegrit/Cell.java
@@ -13,8 +13,8 @@ final class Cell extends Area implements Comparable< Cell> {
 
     //lat, lon format unless specified otherwise; 
     //meant for direct use with elasticsearch Java API FilterBuilder
-    protected long[] topLeft;
-    protected long[] bottomRight;
+    protected double[] topLeft;
+    protected double[] bottomRight;
     
     
     @Override
